split group tests out to own file
